Simplify random ordering code in who-we-are.md
Use the list of people from the DOM instead of duplicating it in the JS.
This might cause an empty list to flash before being replaced but in
practice does not seem noticeable at all. Also remove the <noscript>
because the original HTML will still have the static list before any JS
runs.