Listen to URL fragment changes to update the filter
This means you can edit the URL fragment and the filter will update. Not
a primary mode of usage but if we can support it, we might as well.

This also fixes a few bugs from previous commits in the same PR
involving the interaction between the autocomplete and URL when making
repeated searches. (Basically, we should squash all the commits before
merging into main.)