Persist the county search in the URL across reloads (#43)
* Persist the county search in the URL across reloads

This just puts the county name as a #fragment in the URL and reads it
when the page loads. This will allow the filtered page URL to be
shareable.

* Update filter state from URL only if it matches an existing county

We don't want people sending around URLs with random/malicious filters
in them.

* Listen to URL fragment changes to update the filter

This means you can edit the URL fragment and the filter will update. Not
a primary mode of usage but if we can support it, we might as well.

This also fixes a few bugs from previous commits in the same PR
involving the interaction between the autocomplete and URL when making
repeated searches. (Basically, we should squash all the commits before
merging into main.)

Co-authored-by: Mike Skalnik <mike@skalnik.com>