Capture the exception in catch clause
Old browsers and some automated monitoring tools don't support using a
`catch` clause without a capture variable (this is referred to as
"optional catch binding"). Depending on how we resolve #80 this may be a
moot point, but we are seeing these errors come in via Sentry today.
This should fix WEBSITE-4 (i.e.
https://sentry.io/organizations/vaccinateca/issues/2174390882/).