Add content-hash tagging of assets to bust caches
This is based on an approach at
https://distresssignal.org/busting-css-cache-with-jekyll-md5-hash. I've
modified it to use SHA256 (not for any especially good reason, other
than believing that we shouldn't be writing new code that uses MD5), and
skipped the functionality that looks at directories, since we're not
using any sort of pre-processing on assets currently.

This is minimalistic, but should be effective in cases where clients
have cached old versions of assets (which seems to be the leading cause
of client-reported Sentry errors).