Add linter plugin for new ECMAScript features (#307)
This looks for new ES features that might not be supported by browsers.
I set the cutoff at allowing ES2017 but not ES2018, as that's when
browser support seems to start to get dicey. However, I did manually
whitelist rest spread operators in objects, since that does seem to be
well supported by all browsers released in the last 2 years.

This plugin works best with TypeScript code, where it can tell what the
method receiver is. I've configured it in aggressive mode, which may
increase the risk of false positives. However, without that, it's not
able to catch things like `replaceAll` being called on a string. The
risk of writing code with `replaceAll` (again) outweighs the risk of
false positive to me.

I also had to grab the latest version of eslint-plugin-es from GitHub,
as the most recent release doesn't yet support detecting `replaceAll`,
which sucks since it's, like, the one thing I specifically want.