Switch the JS pipeline to use Webpack
This moves raw JavaScript source files into the `webpack/` directory, to
be transpiled and minified by Webpack + Babel into `assets/js/`, where
they are then picked up by Jekyll as normal.

The `webpack.config.js` file is as minimal as I could manage. It
controls inputs and outputs and generates source maps. In order to
minimize interface changes, the configuration currently generates many
bundles, corresponding to the individual JS files that were previously
directly included. I believe it to be somewhat atypical to have this
many bundles, and we should likely aspire to reduce the number of
separate bundles in the future.

Webpack also generates its own light wrapper in each output file, so
that wrapper is configured to not use modern ("modern") JS features
(like arrow functions). And it is configured to use Babel for actual
processing of the JS source.

Babel is responsible for converting modern JS code into code that is
compatible with older browsers. Its target is configured in the
`.babelrc` file, and is intentionally configured to target extremely old
browsers - our goal is really least common denominator. All of these are
browsers we've observed traffic from (although in many cases, very
little traffic).

As the transpiled JS no longer depends on import/export semantics, we
drop the `type="module"` from the script tags where they're included, as
that's no longer necessary. We also pull in `autoComplete.js` from NPM,
instead of including it as a pre-minified file.

Finally, a new `script/server` makes it easy to run the development
server locally. The development environment uses `node-foreman`, as that
seemed the easiest to manage with our current options for depencency
systems. (We could have used foreman, but Ruby doesn't have the concept
of "dev dependencies" that NPM has)