Stop looking up county for map location
We used to use this to display whether a given site was in your county,
but we no longer show that information. However, we still attempt to do
the county lookup (and block on it before rendering the results). This
should make the site a little faster to render (especially since the
FCC API we're using seems to currently be down).