Embed zip codes and remove opendatasoft fallback
I built a new zipCodes.json file using our [existing ZIP code
database][old ZIP codes] and supplementing with the [database from
opendatasoft][new ZIP codes]. Where data was present in both, I
preferred our existing ZIP code data (as it seems to be more accurate).
I also stripped out everything except the latitude/longitude
coordinates, as we aren't using any other data from the file right now.

This lets us remove the API call to opendatasoft which fired when the
data wasn't in our JSON file (or when the JSON file failed to load).

Additionally, instead of fetching the ZIP code data as a separate JSON
file, we embed it in the webpack bundle, which should eliminate the
possibility of that fetch failing separately from the main JS fetch.

For the time being, I left the original JSON file in place, so that
clients with cached JS (hopefully a dwindling breed....) can continue to
pull the data.

[old ZIP codes]: https://github.com/CAVaccineInventory/gists/blob/main/zip_code_data.json
[new ZIP codes]: https://public.opendatasoft.com/explore/dataset/us-zip-code-latitude-and-longitude/