Lay the groundwork for translated social cards
Today social cards are always rendered in English. This adds the text
description as a translatable string. It also adds the text that's in
the image component of our social cards, and switches to using a
per-lang image. This won't automatically update as the strings are
translated, but once we get translated strings, we can generate new
images.

Goes a long way towards #369 although I would say doesn't fix it until
we get translated text in the images.