Configure Netlify builds with absolute URLs for source maps
Now that JS source is hosted on the Netlify CDN, the relative URLs to
source maps no longer work. We can fix this by configuring an absolute
URL for Netlify builds (which we can put together from environment
variables in the build environment).

Also disable Netlify's JS minification, as that seems to strip the
source map reference.

Fixes #351.