Add "current language" as a translatable string
We're not using this currently, although I'd like to eventually use it
in the language picker (although we need a way through
kurtsson/jekyll-multiple-languages-plugin#159 first). In the mean time,
I'd like to make sure we get a translation from the person that's
translating the rest of the text, just so we don't have to go rooting
around for it ourselves.