Change grid styling on county policy cards
I was wary about adding various conditionals to try and style the grid
items. Instead we're changing the background color to white, and using
grid-gap in order create a separation between each item instead of using
a border and trying to figure out where to put it depending on what
shows or not.