Minor ux improvements color policy cards (#467)
* Change color green to green-600 and green-700

We aren't currently set up to generate CSS as part of our build process
therefore we use Tailwind's colors. From Malthe's comments in issue #377,
he requested they be changed to green-600 and green-700 as they are
closer to correct color (though still not right)

* Fix link overflow in county policy cards

Applied specifically to {{{notes}}} because if it were to be applied to
the links in info URLs et al. the words would split at odd places
depending on screen width.

* Change grid styling on county policy cards

I was wary about adding various conditionals to try and style the grid
items. Instead we're changing the background color to white, and using
grid-gap in order create a separation between each item instead of using
a border and trying to figure out where to put it depending on what
shows or not.

* Change style of grid links on County Policies page