Prep for RTL languages
This adds tailwindcss-rtl, which introduces a bunch of utility classes
meant to handle RTL text better (mostly by referring to "start" and
"end" instead of "left" and "right"). In cases where we are displaying
translated text, the classes that assume LTR text have been switched to
their direction-neutral counterparts. These classes don't work with the
HTML default of dir=auto, so explicitly add dir=ltr to the html tag for
today. In the future, we'll want to derive this from the language.

However, in cases where we are presenting API data that won't be
translated (e.g. detailed report notes), tag those blocks as LTR, as
otherwise formatting of punctuation and such gets wonky. In specific
cases where we're displaying paragraph(s) of text and not snippets,
switch back to explicit left-alignment as well, as is correct for
mixed-mode text.

Even with these changes, map cards seem to be a bit broken. It seems to
behave a bit better if you explicitly pass an RTL language when loading
the Google Maps API
(https://developers.google.com/maps/documentation/javascript/localization#Language);
we may want to consider that in the future.

If you want to test this out (in the absence of an actual Arabic
translation), the easiest way is to edit the top-level dir=ltr attribute
from either your local build or from your browser dev tools.