Add pseudolocales for testing
This introduces the fake locales en-xa and ar-xb (which seem to have
originated with Google, a la
https://developer.android.com/guide/topics/resources/pseudolocales).
en-xa doubles several characters to make the text longer (English is a
comparatively concise language) and adds an assortment of accents to
make sure that UTF-8 is handled correctly. ar-xb is RTL (each string has
Unicode formatting marks at the beginning and end to force RTL handling)
and replaces all letters with upside-down equivalents for ease of
reading.

(ar-xb is a variant of Arabic so that code which has a hard-coded list
of RTL languages will still detect it correctly. Of the two, it's by far
the more useful, since we already have translations into Spanish - a
language with accents and comparatively long phrasing)

There's a script to generate these locales. However, it's a little
overzealous, as it will convert HTML tags, URLs, and Liquid tags that
are embedded in translateable content. I went through and fixed these up
by hand to get something that works.