Pull translations into JS code at build-time (#572)
Instead of dropping translations as an embedded script tag in templates,
this adds a `t` function that JS code can use directly.

I originally intended to use yaml-loader for this, but webpack requires
that dependencies can be statically determined, which would have
required enumerating the language imports explicitly in JS. Instead of
doing that, this adds a new custom loader ("i18n-loader"), which looks
at _config.yml, then iterates over each of the translation files and
dumps them into a single giant object, which it outputs as JS module
source.

The easiest way to review this change is to read
`webpack/loaders/i18n-loader.js` first, then `webpack/i18n.js`. From
there, the rest of the changes are rote substitution using the new `t`
function.

I was mildly worried about the performance costs of not caching the
translation lookups, but when I profiled site card generation, the
translation lookups only seem to amount for about 2.5% of overall
runtime (and if we want to improve overall runtime, I think there's much
lower-hanging fruit).