Remove our use of `carriesVaccine` 😶
This field doesn't really tell us what we want, so lets just ignore it.
We'll show vaccine spotter info as long as we match up, and we'll flip a
no to a yes if we see `carries_vaccine` OR `appointments_available`.
Primarily, this has been an issue with Walgreens locations, which show
`carries_vaccine` as `null`.