Allow users to share links to cards
Change the existing copy behavior to instead construct a URL that autoscrolls to a specific site's
card. This works by constructing a URL with a site ID after `#` when "Share Link" is clicked. After the sites are loaded we check the URL to see if we should autoscroll to a site.