Small page load performance improvement to /near-me (#684)
This defers the work to generate the map info window until after the user actually clicks on the pin in the map. This was showing up as somewhat expensive in a page load profile that I took.

There is a lot of work being done to add a marker to the map for every site, but that seems harder to fix. I just like performance and wanted to see if there were any easy fixes :)