Allow users to share links to cards (#681)
* Allow users to share links to cards

Change the existing copy behavior to instead construct a URL that autoscrolls to a specific site's
card. This works by constructing a URL with a site ID after `#` when "Share Link" is clicked. After the sites are loaded we check the URL to see if we should autoscroll to a site.

* Auto-scroll to sites in county view as well

* Small tweaks to copying links

Links now are always to the near-me page searched by zip code of the site. There's also now a green border around the site when the link is used.

* Record a warning if we're unable to parse the zip code from the address