Create a new /embed endpoint
This endpoint is designed to allow counties to include VaccinateCA data inside of an iframe. It allows customizing the map using new lat, lng, and zoom URL parameters (which also apply to /near-me as well).

Currently this renders all sites on the map that have vaccine available. The list contains all the sites in the map view (no truncation, pagination, or addition of extra sites like /near-me does).

Resolves https://github.com/CAVaccineInventory/site/issues/731