[Embed] Add address search to the embedded map
This switches the /embed site to use google's place autocomplete to enable searching for addresses, zip codes, place names, and more.

I also removed the "Search" button because you now can click on a place name to search for it. There's a bit of ugly error handling. If you type a place that google doesn't know about I fire a JS alert (and this might be frequent if people just type their ZIP codes and hit enter), but I think it's okay for now?