Use the google places API to resolve unknown places
If the user hits ENTER without selecting a place in the autocomplete component, we'll now make a 1-off request to the places autocomplete API to resolve the request. This should eliminate the vast majority of "Unable to find location" errors in the embed page.

From manual testing, it seems to always match the first result in the autocomplete component and only gives an error when the component really has no results.

Annoyingly the autocomplete API doesn't return the lat/lng coordinates so I have to do a subsequent request to get those.